+++ priv/templates/html/confirm_test.exs
@@ -1,0 +1,5 @@
+defmodule <%= base %>.ConfirmTest do
+  use ExUnit.Case
+
+
+end