+++ test/openmaize/custom_auth_test.exs
@@ -1,0 +1,6 @@
+defmodule Openmaize.CustomAuthTest do
+  use ExUnit.Case
+  use Plug.Test
+
+
+end